chore: move actual gate classes into a single file
