+++ quafu/elements/__init__.py
@@ -2,3 +2,4 @@ from .instruction import Instruction, Barrier, Measure, Reset
 from .pulses import Delay, XYResonance, QuantumPulse
 from .quantum_gate import QuantumGate, ControlledGate, MultiQubitGate, SingleQubitGate
 from .classical_element import Cif
+from .unitary import UnitaryDecomposer
