+++ quafu/elements/element_gates/__init__.py
@@ -1,11 +1,4 @@
-from .pauli import *
-from .clifford import HGate, SGate, SdgGate, TGate, TdgGate
-from .rotation import RXGate, RYGate, RZGate, RXXGate, RYYGate, RZZGate, PhaseGate
-from .swap import SwapGate, ISwapGate
-from .c11 import CXGate, CYGate, CZGate, CSGate, CTGate, CPGate
-from .c12 import FredkinGate
-from .cm1 import MCXGate, MCYGate, MCZGate, ToffoliGate
-from .unitary import UnitaryDecomposer
+from .element_gates import *
 
 __all__ = [
     "XGate",
@@ -42,6 +35,5 @@ __all__ = [
     "FredkinGate",
     "MCXGate",
     "MCYGate",
-    "MCZGate",
-    "UnitaryDecomposer",
+    "MCZGate"
 ]
