--- quafu/elements/element_gates/c12.py
@@ -1,10 +1,0 @@
-from ..quantum_gate import ControlledGate, FixedGate, QuantumGate
-from quafu.elements.matrices import SwapMatrix
-
-
-@QuantumGate.register('cswap')
-class FredkinGate(ControlledGate, FixedGate):
-    name = "CSWAP"
-
-    def __init__(self, ctrl: int, targ1: int, targ2: int):
-        ControlledGate.__init__(self, "SWAP", [ctrl], [targ1, targ2], None, tar_matrix=SwapMatrix)
