+++ quafu/elements/element_gates/clifford.py
@@ -1,51 +1,5 @@
-import numpy as np
+from .element_gates import HGate, SGate, CXGate
 
-from quafu.elements.matrices import HMatrix, SMatrix
-from ..quantum_gate import SingleQubitGate, FixedGate, QuantumGate
-
-__all__ = ['HGate', 'SGate', 'SdgGate', 'TGate', 'TdgGate']
+__all__ = ['HGate', 'SGate', 'CXGate']
 
 
-@QuantumGate.register('h')
-class HGate(SingleQubitGate, FixedGate):
-    name = "H"
-    matrix = HMatrix
-
-    def __init__(self, pos: int):
-        FixedGate.__init__(self, pos)
-
-
-@QuantumGate.register('s')
-class SGate(SingleQubitGate, FixedGate):
-    name = "S"
-    matrix = SMatrix
-
-    def __init__(self, pos: int):
-        FixedGate.__init__(self, pos)
-
-
-@QuantumGate.register('sdg')
-class SdgGate(SingleQubitGate, FixedGate):
-    name = "Sdg"
-    matrix = SMatrix.conj().T
-
-    def __init__(self, pos: int):
-        FixedGate.__init__(self, pos)
-
-
-@QuantumGate.register('t')
-class TGate(SingleQubitGate, FixedGate):
-    name = "T"
-    matrix = np.array([[1.0, 0.0], [0.0, np.exp(1.0j * np.pi / 4)]], dtype=complex)
-
-    def __init__(self, pos: int):
-        FixedGate.__init__(self, pos)
-
-
-@QuantumGate.register('tdg')
-class TdgGate(SingleQubitGate, FixedGate):
-    name = "Tdg"
-    matrix = TGate.matrix.conj().T
-
-    def __init__(self, pos: int):
-        FixedGate.__init__(self, pos)
